Add user lang in geocoder query + cleanup (#124)
* Add user lang in geocoder query

* Add test about lang in geocoder, clean tests config

* Dockerfile: Use stretch-slim as base image

* Add comment about puppeteer timeout

+++ tests/integration/tools.js
@@ -37,6 +37,9 @@ export function store(page, soreKey,  data) {
 }
 
 export const clearStore = async function(page) {
+  if(page.url() === 'about:blank') {
+    return
+  }
   await page.evaluate(() =>
     localStorage.clear()
   )
